+++ instrumentation/opentelemetry-instrumentation-asgi/tests/test_asgi_middleware.py
@@ -193,7 +193,7 @@ class TestAsgiApplication(AsgiTestBase):
         self.validate_outputs(outputs, error=ValueError)
 
     def test_override_span_name(self):
-        """Test that span_names can be overwritten by our callback function."""
+        """Test that default span_names can be overwritten by our callback function."""
         span_name = "Dymaxion"
 
         def get_predefined_span_details(_):
@@ -210,7 +210,7 @@ class TestAsgiApplication(AsgiTestBase):
             return expected
 
         app = otel_asgi.OpenTelemetryMiddleware(
-            simple_asgi, span_details_callback=get_predefined_span_details
+            simple_asgi, default_span_details=get_predefined_span_details
         )
         self.seed_app(app)
         self.send_default_request()
@@ -366,6 +366,39 @@ class TestAsgiApplication(AsgiTestBase):
         self.send_default_request()
         span_list = self.memory_exporter.get_finished_spans()
         self.assertEqual(len(span_list), 0)
+
+    def test_hooks(self):
+        def server_request_hook(span, scope):
+            span.update_name("name from server hook")
+
+        def client_request_hook(recieve_span, request):
+            recieve_span.update_name("name from client request hook")
+
+        def client_response_hook(send_span, response):
+            send_span.set_attribute("attr-from-hook", "value")
+
+        def update_expected_hook_results(expected):
+            for entry in expected:
+                if entry["kind"] == trace_api.SpanKind.SERVER:
+                    entry["name"] = "name from server hook"
+                elif entry["name"] == "/ http receive":
+                    entry["name"] = "name from client request hook"
+                elif entry["name"] == "/ http send":
+                    entry["attributes"].update({"attr-from-hook": "value"})
+            return expected
+
+        app = otel_asgi.OpenTelemetryMiddleware(
+            simple_asgi,
+            server_request_hook=server_request_hook,
+            client_request_hook=client_request_hook,
+            client_response_hook=client_response_hook,
+        )
+        self.seed_app(app)
+        self.send_default_request()
+        outputs = self.get_all_output()
+        self.validate_outputs(
+            outputs, modifiers=[update_expected_hook_results]
+        )
 
 
 class TestAsgiAttributes(unittest.TestCase):