+++ instrumentation/opentelemetry-instrumentation-urllib/src/opentelemetry/instrumentation/urllib/__init__.py
@@ -32,12 +32,39 @@ Usage
     req = request.Request('https://postman-echo.com/post', method="POST")
     r = request.urlopen(req)
 
+Hooks
+*******
+
+The urllib instrumentation supports extending tracing behavior with the help of
+request and response hooks. These are functions that are called back by the instrumentation
+right after a Span is created for a request and right before the span is finished processing a response respectively.
+The hooks can be configured as follows:
+
+..code:: python
+
+    # `request_obj` is an instance of urllib.request.Request
+    def request_hook(span, request_obj):
+        pass
+
+    # `request_obj` is an instance of urllib.request.Request
+    # `response` is an instance of http.client.HTTPResponse
+    def response_hook(span, request_obj, response)
+        pass
+
+    URLLibInstrumentor.instrument(
+        request_hook=request_hook, response_hook=response_hook)
+    )
+
 API
 ---
 """
 
 import functools
 import types
+import typing
+
+# from urllib import response
+from http import client
 from typing import Collection
 from urllib.request import (  # pylint: disable=no-name-in-module,import-error
     OpenerDirector,
@@ -54,7 +81,7 @@ from opentelemetry.instrumentation.utils import (
 )
 from opentelemetry.propagate import inject
 from opentelemetry.semconv.trace import SpanAttributes
-from opentelemetry.trace import SpanKind, get_tracer
+from opentelemetry.trace import Span, SpanKind, get_tracer
 from opentelemetry.trace.status import Status
 from opentelemetry.util.http import remove_url_credentials
 
@@ -63,6 +90,11 @@ from opentelemetry.util.http import remove_url_credentials
 _SUPPRESS_HTTP_INSTRUMENTATION_KEY = context.create_key(
     "suppress_http_instrumentation"
 )
+
+_RequestHookT = typing.Optional[typing.Callable[[Span, Request], None]]
+_ResponseHookT = typing.Optional[
+    typing.Callable[[Span, Request, client.HTTPResponse], None]
+]
 
 
 class URLLibInstrumentor(BaseInstrumentor):
@@ -79,18 +111,15 @@ class URLLibInstrumentor(BaseInstrumentor):
         Args:
             **kwargs: Optional arguments
                 ``tracer_provider``: a TracerProvider, defaults to global
-                ``span_callback``: An optional callback invoked before returning the http response.
-                 Invoked with Span and http.client.HTTPResponse
-                ``name_callback``: Callback which calculates a generic span name for an
-                    outgoing HTTP request based on the method and url.
-                    Optional: Defaults to get_default_span_name.
+                ``request_hook``: An optional callback invoked that is invoked right after a span is created.
+                ``response_hook``: An optional callback which is invoked right before the span is finished processing a response
         """
         tracer_provider = kwargs.get("tracer_provider")
         tracer = get_tracer(__name__, __version__, tracer_provider)
         _instrument(
             tracer,
-            span_callback=kwargs.get("span_callback"),
-            name_callback=kwargs.get("name_callback"),
+            request_hook=kwargs.get("request_hook"),
+            response_hook=kwargs.get("response_hook"),
         )
 
     def _uninstrument(self, **kwargs):
@@ -103,12 +132,11 @@ class URLLibInstrumentor(BaseInstrumentor):
         _uninstrument_from(opener, restore_as_bound_func=True)
 
 
-def get_default_span_name(method):
-    """Default implementation for name_callback, returns HTTP {method_name}."""
-    return "HTTP {}".format(method).strip()
-
-
-def _instrument(tracer, span_callback=None, name_callback=None):
+def _instrument(
+    tracer,
+    request_hook: _RequestHookT = None,
+    response_hook: _ResponseHookT = None,
+):
     """Enables tracing of all requests calls that go through
     :code:`urllib.Client._make_request`"""
 
@@ -143,11 +171,7 @@ def _instrument(tracer, span_callback=None, name_callback=None):
         method = request.get_method().upper()
         url = request.full_url
 
-        span_name = ""
-        if name_callback is not None:
-            span_name = name_callback(method, url)
-        if not span_name or not isinstance(span_name, str):
-            span_name = get_default_span_name(method)
+        span_name = "HTTP {}".format(method).strip()
 
         url = remove_url_credentials(url)
 
@@ -160,6 +184,8 @@ def _instrument(tracer, span_callback=None, name_callback=None):
             span_name, kind=SpanKind.CLIENT
         ) as span:
             exception = None
+            if callable(request_hook):
+                request_hook(span, request)
             if span.is_recording():
                 span.set_attribute(SpanAttributes.HTTP_METHOD, method)
                 span.set_attribute(SpanAttributes.HTTP_URL, url)
@@ -193,8 +219,8 @@ def _instrument(tracer, span_callback=None, name_callback=None):
                         ver_[:1], ver_[:-1]
                     )
 
-            if span_callback is not None:
-                span_callback(span, result)
+            if callable(response_hook):
+                response_hook(span, request, result)
 
             if exception is not None:
                 raise exception.with_traceback(exception.__traceback__)